Create mart.transforms.vision naming space for vision-related components.
