+++ mart/transforms/vision/__init__.py
@@ -1,0 +1,12 @@
+#
+# Copyright (C) 2022 Intel Corporation
+#
+# SPDX-License-Identifier: BSD-3-Clause
+#
+
+
+from ...utils.imports import _HAS_TORCHVISION
+
+if _HAS_TORCHVISION:
+    from .objdet import *  # noqa: F403
+    from .transforms import *  # noqa: F403
