+++ mart/transforms/vision/objdet/__init__.py
@@ -1,0 +1,4 @@
+from ....utils.imports import _HAS_PYCOCOTOOLS, _HAS_TORCHVISION
+
+if _HAS_TORCHVISION and _HAS_PYCOCOTOOLS:
+    from .extended import *  # noqa: F403
